IS IT A CAT OR A DOG ?
The final course project was to design and build classifiers that would effectively classify images of cats and dogs using MATLAB. We took twenty-five thousand of images from the Kaggle dataset. Among them, 12,500 were cats, and rest 12,500 were dogs. The motivation behind this project came from Dogs vs. Cats competition from Kaggle which is trying to solve the CAPTCHA (Elson et al., 2007) challenge, which relies on the problem of distinguishing images of dogs and cats. It is easy for humans, but the evidence (Elson et al., 2007) suggests that cats and dogs are particularly difficult to tell apart automatically. A human can accomplish it quickly and easily, but to train an artificial intelligence network is not so easy.
The aim of this project is to effectively use machine learning and deep learning algorithms in the field of image classification problem by using MATLAB. We choose a set of images around 60% of the total dataset as a training set (20,000 images) to train the classifier and used 20% of the total dataset as a validation set (5000 images) to validate the accuracy of the classifier. Finally, rest 20% of the total data used as a test-set (5000 images) to validate the testing accuracy of the classifier using a “confusion matrix.”